# docs/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Raw Data API ",
        "version": "1",
        "description": "Raw Data API  is a set of high-performant APIs for transforming and exporting OpenStreetMap (OSM) data in different GIS file formats." 
    },
...
    "servers": [
        {
            "url": "/latest"
        }
    ],
    "paths": {
...
        "/snapshot/plain": {
            "post": {
                "summary": "Get Current Snapshot As Plain Geojson",
                "description": "Simple API to get osm features as geojson for small region. This is designed only for querying small data for large data follow /snapshot/\n\nParams ::\n\n    bbox: Optional List = takes xmin, ymin, xmax, ymax uses srid=4326\n    select: List = this is select query  you can pass [*] to select all attribute\n    where: List[WhereCondition] = [{'key': 'building', 'value': ['*']},{'key':'amenity','value':['school','college']}]\n    join_by: Optional[JoinFilterType] = or/ and\n    look_in: Optional[List[OsmFeatureType]] = [\"nodes\", \"ways_poly\",\"ways_line\",\"relations\"] : tables name",
                "operationId": "get_current_snapshot_as_plain_geojson_snapshot_plain__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "title": "Params",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/SnapshotParamsPlain"
                                    }
                                ],
                                "default": {}
                            },
                            "examples": {
                                "normal": {
                                    "summary": "Example : Country Boundary",
                                    "description": "**Query** to extract administrative boundary of nepal in plain geojson format",
                                    "value": {
                                        "select": [
                                            "name"
                                        ],
                                        "where": [
                                            {
                                                "key": "admin_level",
                                                "value": [
                                                    "2"
                                                ]
                                            },
                                            {
                                                "key": "boundary",
                                                "value": [
                                                    "administrative"
                                                ]
                                            },
                                            {
                                                "key": "name:en",
                                                "value": [
                                                    "Nepal"
                                                ]
                                            }
                                        ],
                                        "joinBy": "AND",
                                        "lookIn": [
                                            "relations"
                                        ]
                                    }
                                },
                                "second": {
                                    "summary": "Example : City Boundary",
                                    "description": "**Query** to extract city bounadry in plain geojson format",
                                    "value": {
                                        "select": [
                                            "name"
                                        ],
                                        "where": [
                                            {
                                                "key": "admin_level",
                                                "value": [
                                                    "7"
                                                ]
                                            },
                                            {
                                                "key": "boundary",
                                                "value": [
                                                    "administrative"
                                                ]
                                            },
                                            {
                                                "key": "name",
                                                "value": [
                                                    "Pokhara"
                                                ]
                                            }
                                        ],
                                        "joinBy": "AND",
                                        "lookIn": [
                                            "relations"
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Get Current Snapshot As Plain Geojson Snapshot Plain  Post",
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/status/{task_id}": {
            "get": {
                "summary": "Get Task Status",
                "description": "Tracks the request from the task id provided by Raw Data API  for the request\n\nArgs:\n\n    task_id ([type]): [Unique id provided on response from /snapshot/]\n\nReturns:\n\n    id: Id of the task\n    status : SUCCESS / PENDING\n    result : Result of task\n\nSuccessful task will have additional nested json inside",
                "operationId": "get_task_status_tasks_status__task_id___get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Task Id"
                        },
                        "name": "task_id",
                        "in": "path",
                        "description": "Unique id provided on response from /snapshot/" 
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SnapshotTaskResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
...
            "SnapshotTaskResponse": {
                "title": "SnapshotTaskResponse",
                "required": [
                    "id",
                    "status",
                    "result"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string"
                    },
                    "status": {
                        "title": "Status",
                        "type": "string"
                    },
                    "result": {
                        "$ref": "#/components/schemas/SnapshotTaskResult"
                    }
                },
                "additionalProperties": false,
                "example": {
                    "id": "3fded368-456f-4ef4-a1b8-c099a7f77ca4",
                    "status": "SUCCESS",
                    "result": {
                        "download_url": "https://s3.us-east-1.amazonaws.com/exports-stage.hotosm.org/Raw_Export_3fded368-456f-4ef4-a1b8-c099a7f77ca4_GeoJSON.zip",
                        "file_name": "Raw_Export_3fded368-456f-4ef4-a1b8-c099a7f77ca4_GeoJSON",
                        "response_time": "0:00:12.175976",
                        "query_area": "6 Sq Km ",
                        "binded_file_size": "7 MB",
                        "zip_file_size_bytes": 1331601
                    }
                }
            },
            "SnapshotTaskResult": {
                "title": "SnapshotTaskResult",
                "required": [
                    "downloadUrl",
                    "fileName",
                    "responseTime",
                    "queryArea",
                    "bindedFileSize",
                    "zipFileSizeBytes"
                ],
                "type": "object",
                "properties": {
                    "downloadUrl": {
                        "title": "Downloadurl",
                        "type": "string"
                    },
                    "fileName": {
                        "title": "Filename",
                        "type": "string"
                    },
                    "responseTime": {
                        "title": "Responsetime",
                        "type": "string"
                    },
                    "queryArea": {
                        "title": "Queryarea",
                        "type": "string"
                    },
                    "bindedFileSize": {
                        "title": "Bindedfilesize",
                        "type": "string"
                    },
                    "zipFileSizeBytes": {
                        "title": "Zipfilesizebytes",
                        "type": "integer"
                    }
                },
                "additionalProperties": false
            },
...
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            },
...
        }
    }
}